week3-ztna/idp/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
PyJWT>=2.9.0
pydantic>=2.10.0
